cliques in a tidy format
This function extracts in a tiddy format the cliques present in a network per row.
The size of the cliques (i.e. 3, 4, 5, ...) are specified with an integer as the variable of the function.
The outcome is a tibble with as much columns as the size of the clique size and an ID per clique.